Add PlantUML UML output
